requirements.txt
datasketch==1.6.5
faiss-cpu==1.10.0
hdbscan==0.8.40
hnswlib==0.8.0
iteration_utilities==0.13.0
loguru==0.7.3
matplotlib==3.10.0
mlflow==2.20.1
munkres==1.1.4
numpy==1.26.4
pandas==2.2.3
pytorch-lightning==1.9.5
scikit-learn==1.6.1
spacy==3.8.3
tqdm==4.67.1
transformers==4.48.3
ujson==5.10.0
umap-learn==0.5.7
xgboost==2.1.4
jupyter==1.1.1
